requirements.txt
# Django Framework
Django==6.0.2
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3

# Database
psycopg2-binary==2.9.11

# Admin Interface
django-jazzmin==3.0.2

# Image Processing
Pillow==12.1.1

# HTTP Requests
requests==2.32.5
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
